Created common layout component for floating box in centre of screen
Used it for pages: profile, layout and login.
